echo_translate/requirements.txt
streamlit
SpeechRecognition
edge-tts
fpdf2
Pillow
requests
streamlit-mic-recorder
